requirements.txt
aiohttp>=3.8.0
python-dotenv>=1.0.0
openai>=1.3.0
typing-extensions>=4.5.0
loguru==0.7.2
pipecat
